Move event system from object to separate event manager (from which EditorApp derives)
Remove the O(n) process of checking N lists of events to a O(1) process of checking a master list of events. Why did I even think that other way was better? This is much more sensible. Event*s are paired with a target Object* on which they are Event::run(...).